WordSearch program
This program takes in a two-dimensional grid of characters and allows a
user to search for words in the grid. The user types in the word they
would like to search for in the grid and the program outputs the
location and orientation of the word in the grid (if the word is in the
grid) or tells the user that the word is not in the grid if it is not
found. The driver for this program was written by my professor last
year (this was a school assignment). Our task was to write the rest of
the code that would do the searching and other necessary methods. I
wrote the WordSearch.java file.